meetingmind/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
box-sdk-gen==1.3.0
python-dotenv==1.0.0